+++ test/spec/raster-service-tests.js
@@ -17,12 +17,6 @@ describe('Testing raster service', function () {
     RasterService = $injector.get('RasterService');
   }));
 
-  it('should return the given value on set/getIntensityData', function () {
-    var thisun = 'intensityData';
-    RasterService.setIntensityData(thisun);
-    expect(RasterService.getIntensityData()).toEqual(thisun);
-  });
-
   it('should return a CabinetService get promise', function () {
     var geom = new L.LatLng(52.50995268098114, 4.961357116699219);
     var result = RasterService.getData({'layer': 'layer'}, {'geom': geom});